+++ anyblok/column.py
@@ -3,10 +3,12 @@
 #    Copyright (C) 2016 Jean-Sebastien SUZANNE <jssuzanne@anybox.fr>
 #    Copyright (C) 2017 Jean-Sebastien SUZANNE <jssuzanne@anybox.fr>
 #    Copyright (C) 2018 Jean-Sebastien SUZANNE <jssuzanne@anybox.fr>
+#    Copyright (C) 2019 Jean-Sebastien SUZANNE <js.suzanne@gmail.com>
 #
 # This Source Code Form is subject to the terms of the Mozilla Public License,
 # v. 2.0. If a copy of the MPL was not distributed with this file,You can
 # obtain one at http://mozilla.org/MPL/2.0/.
+from base64 import b64encode, b64decode
 from .field import Field, FieldException
 from .mapper import ModelAttributeAdapter
 from sqlalchemy.schema import Sequence as SA_Sequence, Column as SA_Column
@@ -19,10 +21,13 @@ from sqlalchemy_utils.types.url import URLType
 from sqlalchemy_utils.types.phone_number import PhoneNumberType
 from sqlalchemy_utils.types.email import EmailType
 from sqlalchemy_utils.types.scalar_coercible import ScalarCoercible
-from datetime import datetime, date
+from sqlalchemy_utils import JSONType
+from datetime import datetime, date, timedelta
 from dateutil.parser import parse
 from inspect import ismethod
 from anyblok.config import Configuration
+from .common import sgdb_in
+from json import dumps, loads
 import time
 import pytz
 import decimal
@@ -167,12 +172,12 @@ class Column(Field):
         ('is crypted', False),
     ))
 
-    def native_type(cls):
+    def native_type(self, registry):
         """Return the native SqlAlchemy type
 
         :rtype: sqlalchemy native type
         """
-        return cls.sqlalchemy_type
+        return self.sqlalchemy_type
 
     def format_foreign_key(self, registry, args, kwargs):
         """Format a foreign key
@@ -226,7 +231,7 @@ class Column(Field):
             else:
                 kwargs['default'] = self.default_val
 
-        sqlalchemy_type = self.sqlalchemy_type
+        sqlalchemy_type = self.native_type(registry)
         if self.encrypt_key:
             encrypt_key = self.format_encrypt_key(registry, namespace)
             sqlalchemy_type = EncryptedType(sqlalchemy_type, encrypt_key)
@@ -352,6 +357,13 @@ class Float(Column):
     sqlalchemy_type = types.Float
 
 
+"""
+    Added *process_result_value* at the class *DECIMAL*, because
+    this method is necessary for encrypt the column
+"""
+types.DECIMAL.process_result_value = lambda self, value, dialect: value
+
+
 class Decimal(Column):
     """Decimal column
 
@@ -377,8 +389,19 @@ class Decimal(Column):
         :return: decimal value
         """
         if value is not None:
-            if not isinstance(value, decimal.Decimal):
+            if self.encrypt_key:
+                value = str(value)
+            elif not isinstance(value, decimal.Decimal):
                 value = decimal.Decimal(value)
+
+        return value
+
+    def getter_format_value(self, value):
+        if value is None:
+            return None
+
+        if self.encrypt_key:
+            value = decimal.Decimal(value)
 
         return value
 
@@ -516,6 +539,31 @@ class DateTime(Column):
         return res
 
 
+class TimeStamp(DateTime):
+    """ TimeStamp column
+
+    ::
+
+        from datetime import datetime
+        from anyblok.declarations import Declarations
+        from anyblok.column import DateTime
+
+
+        @Declarations.register(Declarations.Model)
+        class Test:
+
+            x = TimeStamp(default=datetime.now)
+
+    """
+
+    def __init__(self, *args, **kwargs):
+        super(TimeStamp, self).__init__(*args, **kwargs)
+        self.sqlalchemy_type = types.TIMESTAMP(timezone=True)
+
+    def getter_format_value(self, value):
+        return convert_string_to_datetime(value)
+
+
 class Time(Column):
     """Time column
 
@@ -552,6 +600,26 @@ class Interval(Column):
 
     """
     sqlalchemy_type = types.Interval
+
+    def native_type(self, registry):
+        if self.encrypt_key:
+            return types.VARCHAR(1024)
+
+        return self.sqlalchemy_type
+
+    def setter_format_value(self, value):
+        if self.encrypt_key:
+            value = dumps({
+                x: getattr(value, x)
+                for x in ['days', 'seconds', 'microseconds']})
+
+        return value
+
+    def getter_format_value(self, value):
+        if self.encrypt_key:
+            value = timedelta(**loads(value))
+
+        return value
 
 
 class StringType(types.TypeDecorator):
@@ -632,7 +700,7 @@ class Password(Column):
             ==> 0
     """
     def __init__(self, *args, **kwargs):
-        size = kwargs.pop('size', 64)
+        self.size = kwargs.pop('size', 64)
         crypt_context = kwargs.pop('crypt_context', {})
         self.crypt_context = crypt_context
         kwargs.pop('type_', None)
@@ -640,7 +708,8 @@ class Password(Column):
         if 'foreign_key' in kwargs:
             raise FieldException("Column Password can not have a foreign key")
 
-        self.sqlalchemy_type = PasswordType(max_length=size, **crypt_context)
+        self.sqlalchemy_type = PasswordType(
+            max_length=self.size, **crypt_context)
         super(Password, self).__init__(*args, **kwargs)
 
     def setter_format_value(self, value):
@@ -659,6 +728,7 @@ class Password(Column):
         :return: autodoc properties
         """
         res = super(Password, self).autodoc_get_properties()
+        res['size'] = self.size
         res['Crypt context'] = self.crypt_context
         return res
 
@@ -896,6 +966,11 @@ class Selection(Column):
         :param Model:
         :return: list of checkConstraint
         """
+        if self.encrypt_key:
+            # dont add constraint because the state is crypted and nobody
+            # can add new entry
+            return []
+
         selections = self.sqlalchemy_type.selections
         if isinstance(selections, dict):
             enum = selections.keys()
@@ -925,6 +1000,13 @@ class Selection(Column):
         return []
 
 
+"""
+    Added *process_result_value* at the class *JSON*, because
+    this method is necessary for encrypt the column
+"""
+types.JSON.process_result_value = lambda self, value, dialect: value
+
+
 class Json(Column):
     """JSON column
 
@@ -941,6 +1023,28 @@ class Json(Column):
 
     """
     sqlalchemy_type = types.JSON(none_as_null=True)
+
+    def native_type(self, registry):
+        """ Return the native SqlAlchemy type """
+        if sgdb_in(registry.engine, ['MariaDB']):
+            return JSONType
+
+        return self.sqlalchemy_type
+
+    def setter_format_value(self, value):
+        if self.encrypt_key:
+            value = dumps(value)
+
+        return value
+
+    def getter_format_value(self, value):
+        if value is None:
+            return None
+
+        if self.encrypt_key:
+            value = loads(value)
+
+        return value
 
 
 class LargeBinary(Column):
@@ -963,6 +1067,24 @@ class LargeBinary(Column):
 
     """
     sqlalchemy_type = types.LargeBinary
+
+    def native_type(self, registry):
+        if self.encrypt_key:
+            return types.Text
+
+        return self.sqlalchemy_type
+
+    def setter_format_value(self, value):
+        if self.encrypt_key:
+            value = b64encode(value).decode('utf-8')
+
+        return value
+
+    def getter_format_value(self, value):
+        if self.encrypt_key:
+            value = b64decode(value.encode('utf-8'))
+
+        return value
 
 
 class Sequence(String):
@@ -1321,6 +1443,11 @@ class Country(Column):
 
         :return: list of checkConstraint
         """
+        if self.encrypt_key:
+            # dont add constraint because the state is crypted and nobody
+            # can add new entry
+            return []
+          
         enum = [country.alpha_3 for country in pycountry.countries]
         constraint = """"%s" in ('%s')""" % (self.fieldname, "', '".join(enum))
         enum.sort()