+++ anyblok/field.py
@@ -165,7 +165,7 @@ class Field:
             label = fieldname.replace('_', ' ')
             self.label = label.capitalize()
 
-    def native_type(self):
+    def native_type(self, registry):
         """ Return the native SqlAlchemy type
 
         :exception: FieldException